feat: 체스방 삭제 시 게임이 진행중이면 삭제하지 못하도록 함

+++ src/main/java/chess/web/service/ChessService.java
@@ -133,9 +133,25 @@ public class ChessService {
         return roomDao.findAll();
     }
 
-    public void removeRoom(Long boardId, String password) {
+    public boolean removeRoom(Long boardId, String password) {
+        if (!hasRoom(boardId) || isRunningChess(boardId)) {
+            return false;
+        }
         pieceDao.deleteByBoardId(boardId);
         roomDao.delete(boardId, password);
         boardDao.deleteById(boardId);
+        return true;
+    }
+
+    private boolean hasRoom(Long boardId) {
+        return roomDao.findByBoardId(boardId).isPresent();
+    }
+
+    private boolean isRunningChess(Long boardId) {
+        long kingCount = pieceDao.findAllByBoardId(boardId).stream()
+                .filter(Piece::isKing)
+                .count();
+
+        return kingCount == 2;
     }
 }
